requirements.txt
tqdm
torch
torchvision
pydantic
pillow
numpy
scipy
matplotlib
plotly
POT #ot
